Refactoring: Rename 'InsertContext'
