+++ src/main/java/com/honeypot/common/validation/groups/InsertContext.java
@@ -1,4 +1,4 @@
-package com.honeypot.common.validation;
+package com.honeypot.common.validation.groups;
 
 import javax.validation.groups.Default;
